refactor: reorganize js into src
