+++ index.html
@@ -327,7 +327,7 @@
 
     <script src="https://3Dmol.org/build/3Dmol-min.js"></script>
     <script src="https://unpkg.com/smiles-drawer@2.0.1/dist/smiles-drawer.min.js"></script>
-    <script type="module" src="main.js"></script>
+    <script type="module" src="./src/main.js"></script>
 
     <!-- Quick 3D View Modal -->
     <div id="quick-view-modal" class="modal">
