+++ src/main.js
@@ -1,11 +1,11 @@
-import MoleculeLoader from './services/MoleculeLoader.js';
-import MoleculeRepository from './services/MoleculeRepository.js';
-import BoundLigandTable from './ui/BoundLigandTable.js';
-import FragmentLibrary from './ui/FragmentLibrary.js';
-import LigandModal from './ui/ligandModal.js';
-import MoleculeCard from './ui/MoleculeCard.js';
-import PdbDetailsModal from './ui/PdbDetailsModal.js';
-import ProteinBrowser from './ui/ProteinBrowser.js';
+import MoleculeLoader from './utils/MoleculeLoader.js';
+import MoleculeRepository from './utils/MoleculeRepository.js';
+import BoundLigandTable from './components/BoundLigandTable.js';
+import FragmentLibrary from './components/FragmentLibrary.js';
+import LigandModal from './modal/ligandModal.js';
+import MoleculeCard from './components/MoleculeCard.js';
+import PdbDetailsModal from './modal/PdbDetailsModal.js';
+import ProteinBrowser from './components/ProteinBrowser.js';
 
 class MoleculeManager {
     constructor() {
